// TS_Angular/HW7MyPipe/src/app/models/user.ts
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: UserAddress;
  phone: string;
  website: string;
  company: Company;
}

interface UserAddress {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: {
    lat: number;
    lng: number;
  };
}

interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}